requirements.txt
tgtg==0.17.0
python-telegram-bot==21.6
